Finally found a combination of API endpoint (iTunes) and DRY code that works and I'm happy with.
The API proved really frustrating. Most APIs ask for an API Key, and/or
full authentication. That meant signing up, and in the case of
authentication, it meant learning how to do that. I didn't feel I had
the time (although it took so long to get around the issue I almost feel
I SHOULD have learned that stuff).

Almost all the public APIs I found that don't need authenticating lack
the correct CORS headers. The most frustrating of these were the ones
from Data.gov.uk (which technically I own a share in because I pay tax).

Finally I found the iTunes API, which is slow and awkward, but does give
results.

Making the code DRY was more a process of attrition. I tried a few
different approaches. I've finally settled on one that is really short
and I'm quite pleased with.