Song fields clear when clicking "refresh" button.
This works by merging the click stream into the suggestion stream. In
this way we get a stream that contains either song data or clicks,
indicated by `null`. When the renderer receives `null`, it chooses to
show an appropriate display.

I find this a bit counter-intuitive. This tutorial (and I suppose the
bulk of FRP) seems to be more about encoding behaviours into patterns of
data, and then having the system (often views) respond to changes or
progressions along that stream of data.

Paraphrasing the tutorial:
the song/suggestion fields must show an empty state, a song (on load),
then an empty state (after clicks on refresh), then a song, then empty
then song etc.. with successive clicks. So the data behind the view must
be "empty... song... empty... song... etc", so to drive that flow of
data, the stream must come from "onload... response... click...
response", so we need to create streams that reflect these inputs; the
streams can be separate and then merged.