A cras implementation of reading a text box and submitting the value as a search term to a remote service, using Rx.
It's cras because the keyboard handling is unsubtle. Ideally the system
would respond to any change in text (including deletions and edits), but
that's pretty rich and complex to handle because the keyboard events are
rich and complex.

I only really wanted to see how I could combine some streams to trigger
a request to a url. I think this covers that.

In my implementation, I debounce keypress events on the input to 500ms,
pluck the value of the input from the event, filter out any cases less
than 3 characters, and then map to a URL reflecting a search on the
iTunes content. I'm sure I've seen examples that debounce closer to the
point the strea is joined with any other control streams, but that stuff
confuses me.